Call the correct scripts. To be removed later.
The CI configs are calling the scripts with underscores.
So we need to keep them around until CI scripts get updated. We can't
update the CI scripts without these changes being in the repo.

Plan is get this PR merged. Then post an update to CI scripts. Then post
a follow up to this repo to remove these underscored scripts.